Remove dep on http-binding from client-http
And also moves rest-json into its own package so that the awsjson
protocols don't have to pull in http bindings.
