+++ aws/client-awsjson/src/it/java/software/amazon/smithy/java/runtime/client/aws/jsonprotocols/AwsJson1ProtocolTests.java
