+++ aws/client-awsjson/src/main/resources/META-INF/services/software.amazon.smithy.java.runtime.client.core.ClientProtocolFactory
@@ -1,0 +1,2 @@
+software.amazon.smithy.java.runtime.aws.client.awsjson.AwsJson1Protocol$Factory
+software.amazon.smithy.java.runtime.aws.client.awsjson.AwsJson11Protocol$Factory
