--- aws/client-json-protocols/src/main/resources/META-INF/services/software.amazon.smithy.java.runtime.client.core.ClientProtocolFactory
@@ -1,3 +1,0 @@
-software.amazon.smithy.java.runtime.client.aws.jsonprotocols.AwsJson1Protocol$Factory
-software.amazon.smithy.java.runtime.client.aws.jsonprotocols.AwsJson11Protocol$Factory
-software.amazon.smithy.java.runtime.client.aws.jsonprotocols.RestJsonClientProtocol$Factory
