+++ aws/client-restjson/src/main/resources/META-INF/services/software.amazon.smithy.java.runtime.client.core.ClientProtocolFactory
@@ -1,0 +1,1 @@
+software.amazon.smithy.java.runtime.aws.client.restjson.RestJsonClientProtocol$Factory
