+++ codegen/client/src/it/java/software/amazon/smithy/java/codegen/client/AuthSchemeTest.java
@@ -15,8 +15,8 @@ import org.junit.jupiter.api.Test;
 import smithy.java.codegen.server.test.client.TestServiceClient;
 import smithy.java.codegen.server.test.model.EchoInput;
 import software.amazon.smithy.java.codegen.client.util.EchoServer;
+import software.amazon.smithy.java.runtime.aws.client.restjson.RestJsonClientProtocol;
 import software.amazon.smithy.java.runtime.client.auth.api.scheme.AuthSchemeOption;
-import software.amazon.smithy.java.runtime.client.aws.jsonprotocols.RestJsonClientProtocol;
 import software.amazon.smithy.java.runtime.client.core.interceptors.ClientInterceptor;
 import software.amazon.smithy.java.runtime.client.core.interceptors.RequestHook;
 import software.amazon.smithy.java.runtime.core.schema.PreludeSchemas;
