+++ settings.gradle.kts
@@ -26,6 +26,7 @@ include(":json-codec")
 include(":client-core")
 include(":client-endpoint-api")
 include(":client-http")
+include(":client-http-binding")
 include(":client-auth-api")
 
 include(":auth-api")
@@ -46,4 +47,5 @@ include("server-core")
 // AWS specific
 include(":aws:event-streams")
 include(":aws:client-http")
-include(":aws:client-json-protocols")
+include(":aws:client-awsjson")
+include(":aws:client-restjson")
